requirements.txt
# =============================================================================
# Rice Leaf Disease Classifier — Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# Deep Learning
tensorflow>=2.12.0
keras-tuner>=1.3.0

# Web Framework (API + server)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6      # Required for file uploads in FastAPI

# Data Processing & ML
numpy>=1.23.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0                # Statistical tests (McNemar, Wilcoxon, t-test)

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0         # Grad-CAM heatmap overlays

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
PyYAML>=6.0                  # Config file loading
tqdm>=4.65.0                 # Progress bars

# Deployment (Hugging Face Spaces)
gradio>=4.0.0                # Optional: simple demo UI for HF Spaces
